collect common Connection args into ConnectionOpts
Positional arguments are mandatory, and members of the opts struct are
optional, which makes it much easier for us to add new optional
paramters. The following commit will add a parameter to control the
connection backoff, for example. This commit is just a reshuffling of
parameters without any behavior change, so it should be easy to
integrate in client/kbfs/gregor.